63 product domain

+++ features/products/infrastructure/product_supabase_data.ts
@@ -1,6 +1,7 @@
 import { SupabaseClient } from '@supabase/supabase-js'
 import { Database } from 'backend/database.types'
-import { LimitIsNotInRange } from '../../shared/infrastructure/limit_is_not_in_range'
+import { ParameterNotMatchException } from '../../shared/infrastructure/parameter_not_match_exception'
+import { LimitIsNotInRangeException } from '../../shared/infrastructure/limit_is_not_in_range_exception'
 import { KeyAlreadyExistException } from '../../shared/infrastructure/key_already_exist_exception'
 import { ValidRank } from '../../shared/domain/value_objects/ValidRank'
 import { BaseException } from '../../shared/domain/exceptions/BaseException'
@@ -20,20 +21,6 @@ export class ProductSupabaseData implements ProductRepository {
 
 	readonly tableName = 'products'
 
-	async deleteAll(): Promise<boolean> {
-		try {
-			await this.client.from( this.tableName )
-			          .delete()
-
-			return true
-		}
-		catch ( e ) {
-			console.log( 'supabase unexpected error' )
-			console.log( e )
-			throw [ new InfrastructureException() ]
-		}
-	}
-
 	async createProduct( product: Product ): Promise<boolean> {
 		const result = await this.client.from( this.tableName )
 		                         .insert( productToJson( product ) as any )
@@ -47,21 +34,31 @@ export class ProductSupabaseData implements ProductRepository {
 		return true
 	}
 
-	async getAll( from: ValidInteger, to: ValidInteger ): Promise<Product[]> {
+	async getAll( from: ValidInteger, to: ValidInteger,
+		name?: ValidString ): Promise<Product[]> {
 
-		const result = await this.client.from( this.tableName )
-		                         .select()
-		                         .range( from.value, to.value )
+		const result = this.client.from( this.tableName )
+		                   .select()
 
-		if ( result.error ) {
-			if ( result.error.code === 'PGRST103' ) {
-				throw [ new LimitIsNotInRange() ]
+		if ( name !== undefined ) {
+			result.eq( 'name', name.value )
+		}
+
+		const { data, error } = await result.range( from.value, to.value )
+
+		if ( name != undefined && data?.length === 0 ) {
+			throw [ new ParameterNotMatchException( 'name' ) ]
+		}
+
+		if ( error ) {
+			if ( error.code === 'PGRST103' ) {
+				throw [ new LimitIsNotInRangeException() ]
 			}
 			throw [ new InfrastructureException() ]
 		}
 
 		const products: Product[] = []
-		for ( const json of result.data ) {
+		for ( const json of data ) {
 
 			const product = productFromJson( json )
 
@@ -75,48 +72,60 @@ export class ProductSupabaseData implements ProductRepository {
 	}
 
 	async getProduct( code: ValidString ): Promise<Product> {
-		const result = await this.client.from( this.tableName )
-		                         .select()
-		                         .eq( 'product_code', code.value )
+		try {
+			const result = await this.client.from( this.tableName )
+			                         .select()
+			                         .eq( 'product_code', code.value )
 
-		if ( result.error ) {
-			console.log( 'supabase unexpected error' )
-			console.log( result.error )
-			throw [ new InfrastructureException() ]
+			if ( result.error ) {
+				throw [ new InfrastructureException() ]
+			}
+
+			if ( result.data.length === 0 ) {
+				throw [ new ParameterNotMatchException( 'product_code' ) ]
+			}
+
+			const product = productFromJson( result.data[0] )
+
+			if ( product instanceof BaseException ) {
+				throw product
+			}
+
+			return product as Product
 		}
-
-		const product = productFromJson( result.data[0] )
-
-		if ( product instanceof BaseException ) {
-			throw product
+		catch ( e ) {
+			throw e
 		}
-
-		return product as Product
 	}
 
 	async updateProduct(
+		product_code: ValidString,
 		product: Product
 	): Promise<boolean> {
 		try {
-			const result = await this.client.from( this.tableName )
+			await this.client.from( this.tableName )
 			          .update( productToJson( product ) as any )
 			          .eq(
 				          'product_code',
-				          product.product_code.value
+				          product_code.value
 			          )
-			console.log("result")
-			console.log(result)
 			return true
 		}
 		catch ( e ) {
-			console.log( 'supabase unexpected error' )
-			console.log( e )
 			throw [ new InfrastructureException() ]
 		}
 	}
 
 	async deleteProduct( code: ValidString ): Promise<boolean> {
 		try {
+			const result = await this.client.from( this.tableName )
+			                         .select()
+			                         .eq( 'product_code', code.value )
+
+			if ( result.data?.length === 0 ) {
+				throw [ new ParameterNotMatchException( 'product_code' ) ]
+			}
+
 			await this.client.from( this.tableName )
 			          .delete()
 			          .eq(
@@ -126,42 +135,8 @@ export class ProductSupabaseData implements ProductRepository {
 			return true
 		}
 		catch ( e ) {
-			console.log( 'supabase unexpected error' )
-			console.log( e )
-			throw [ new InfrastructureException() ]
+			throw e
 		}
-	}
-
-	async searchProduct( name: ValidString, from: ValidInteger,
-		to: ValidInteger ): Promise<Product[]> {
-
-		const result = await this.client.from( this.tableName )
-		                         .select()
-		                         .range( from.value, to.value )
-		                         .ilike( 'name', `%${ name.value }%` )
-
-		if ( result.error ) {
-			console.log( 'supabase unexpected error' )
-			console.log( result.error )
-			throw [ new InfrastructureException() ]
-		}
-
-		const errors: BaseException[] = []
-
-		const products: Product[] = []
-
-		for ( const json of result.data ) {
-			const product = productFromJson( json )
-
-			if ( product instanceof BaseException ) {
-				errors.push( product )
-				throw errors
-			}
-
-			products.push( product as Product )
-		}
-
-		return products
 	}
 
 	public getRecommendProductsGroupByCategory( threshold: ValidRank,
